chore: merge app connection flows
